# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://api.revmetrix.io"
    }
  ],
  "paths": {
...
    "/api/gets/GetAcclerationFromBall_id": {
      "get": {
        "tags": [
          "Gets"
        ],
        "operationId": "GetAcclerationFromBall_id",
        "parameters": [
          {
            "name": "ball_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shot"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shot"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shot"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/gets/GetStrikesVsShotsPercentage": {
      "get": {
        "tags": [
          "Gets"
        ],
        "operationId": "GetStrikesVsShotsPercentage",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shot"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shot"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Shot"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/gets/GetUsers": {
      "get": {
        "tags": [
          "Gets"
        ],
        "operationId": "GetUsers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserIdentification"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserIdentification"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserIdentification"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/Authorize": {
      "post": {
        "tags": [
          "Posts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/Register": {
      "post": {
        "tags": [
          "Posts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserIdentification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/posts/Refresh": {
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "Refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ByteArrayToken"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ByteArrayToken"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ByteArrayToken"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DualToken"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/posts/HashAndSalt": {
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "HashAndSalt",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/HashAndSalt"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashAndSalt"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/HashAndSalt"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ByteArrayToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DualToken": {
        "type": "object",
        "properties": {
          "tokenA": {
            "type": "string",
            "nullable": true
          },
          "tokenB": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HashAndSalt": {
        "type": "object",
        "properties": {
          "hash": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "salt": {
            "type": "string",
            "format": "byte",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Password": {
        "type": "object",
        "properties": {
          "rawPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "Shot": {
        "type": "object",
        "properties": {
          "shot_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "session_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "game_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "frame_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ball_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "video_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "shot_number": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "shot_number_ot": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "lane_Number": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "pocket_hit": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "count": {
            "type": "string",
            "nullable": true
          },
          "pins": {
            "type": "string",
            "nullable": true
          },
          "ddx": {
            "type": "number",
            "format": "float"
          },
          "ddy": {
            "type": "number",
            "format": "float"
          },
          "ddz": {
            "type": "number",
            "format": "float"
          },
          "x_position": {
            "type": "number",
            "format": "float"
          },
          "y_position": {
            "type": "number",
            "format": "float"
          },
          "z_position": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
...
      "UserIdentification": {
        "type": "object",
        "properties": {
          "firstname": {
            "type": "string",
            "nullable": true
          },
          "lastname": {
            "type": "string",
            "nullable": true
          },
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "phone_number": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}